Revert "feat(custom-button): slit custom button by desktop / mobile"
This reverts commit 63a1c891b978d8fdde0f43e9083d17e385b16efc.
