--- .changeset/tall-mugs-clap.md
@@ -1,5 +1,0 @@
----
-'@alfalab/core-components-custom-button': patch
----
-
-Добавлена возможность импортировать desktop / mobile по отдельности
