+++ packages/custom-button/src/Component.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 
-import { CustomButton, CustomButtonDesktop, CustomButtonMobile } from './index';
+import { CustomButton } from './index';
 
 const dataTestId = 'test-id';
 
@@ -23,31 +23,21 @@ describe('CustomButton', () => {
     describe('Snapshots tests', () => {
         it('should match snapshot', () => {
             expect(render(<CustomButton />)).toMatchSnapshot();
-            expect(render(<CustomButtonDesktop />)).toMatchSnapshot();
-            expect(render(<CustomButtonMobile />)).toMatchSnapshot();
         });
 
         it('should render custom background color', () => {
             expect(render(<CustomButton backgroundColor='#00ff00' />)).toMatchSnapshot();
-            expect(render(<CustomButtonDesktop backgroundColor='#00ff00' />)).toMatchSnapshot();
-            expect(render(<CustomButtonMobile backgroundColor='#00ff00' />)).toMatchSnapshot();
         });
 
         it('should render black color content', () => {
             expect(render(<CustomButton contentColor='black' />)).toMatchSnapshot();
-            expect(render(<CustomButtonDesktop contentColor='black' />)).toMatchSnapshot();
-            expect(render(<CustomButtonMobile contentColor='black' />)).toMatchSnapshot();
         });
 
         it('should render CustomButton view=primary by default', () => {
             expect(render(<CustomButton />)).toMatchSnapshot();
-            expect(render(<CustomButtonDesktop />)).toMatchSnapshot();
-            expect(render(<CustomButtonMobile />)).toMatchSnapshot();
         });
         it('should render loader if loading pass', () => {
             expect(render(<CustomButton loading={true} />)).toMatchSnapshot();
-            expect(render(<CustomButtonDesktop loading={true} />)).toMatchSnapshot();
-            expect(render(<CustomButtonMobile loading={true} />)).toMatchSnapshot();
         });
     });
