--- packages/custom-button/src/constants/default-colors.ts
@@ -1,2 +1,0 @@
-export const DEFAULT_BUTTON_COLOR = '#FF45C3';
-export const DEFAULT_CONTENT_COLOR = 'white';
