--- packages/custom-button/src/desktop/index.ts
@@ -1,1 +1,0 @@
-export { CustomButtonDesktop } from './Component.desktop';
