+++ packages/custom-button/src/index.ts
@@ -1,5 +1,1 @@
-export { CustomButton } from './Component';
-export { CustomButtonProps, ComponentProps } from './types/props';
-
-export { CustomButtonDesktop } from './desktop';
-export { CustomButtonMobile } from './mobile';
+export * from './Component';
