--- packages/custom-button/src/mobile/index.ts
@@ -1,1 +1,0 @@
-export { CustomButtonMobile } from './Component.mobile';
